# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: IDP Perimeter API
  version: 0.3.0
  description: >
    Контракт периметра (портал ↔ API-шлюз) платформы IDP. Источник правды для
    кодогена TS-клиента и zod-схем.
servers:
  - url: /api
paths:
...
  /iam/roles:
    get:
      operationId: listRoles
      summary: Список ролей каталога RBAC
      description: >
        Возвращает все роли каталога (read-only). Привилегированная ручка
        IAM-админки: требует право (read, iam:global); отказ/недоступность IDM →
        403 (fail-closed). Роли/права сидируются миграциями — UI их только
        показывает (ADR-0014).
      responses:
        "200":
          description: Список ролей
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleList"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createRole
      summary: Создать роль каталога RBAC
      description: >
        Создаёт пользовательскую роль (system=false). Привилегированная
        структурная мутация каталога: требует право (manage, iam:global);
        отказ/недоступность IDM → 403 (fail-closed). Повторное создание роли с тем
        же именем → 409 (не идемпотентно); пустое имя → 400. После мутации IDM
        широко инвалидирует кэш решений (ADR-0015).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoleRequest"
      responses:
        "201":
          description: Роль создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /iam/roles/{role}:
    delete:
      operationId: deleteRole
      summary: Удалить роль каталога RBAC
      description: >
        Удаляет роль (каскадно снимает её у всех носителей и убирает связки прав).
        Требует право (manage, iam:global); отказ/недоступность IDM → 403
        (fail-closed). Системная (сидированная) роль защищена от удаления → 422;
        несуществующая роль → 404. После мутации IDM широко инвалидирует кэш
        решений, т.к. затронуты все носители роли (ADR-0015).
      parameters:
        - $ref: "#/components/parameters/RolePath"
      responses:
        "200":
          description: Роль удалена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/PreconditionFailed"
  /iam/permissions:
    get:
      operationId: listPermissions
      summary: Список прав каталога RBAC
      description: >
        Возвращает все права каталога (read-only). Требует право (read,
        iam:global); отказ/недоступность IDM → 403 (fail-closed).
      responses:
        "200":
          description: Список прав
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermissionList"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createPermission
      summary: Создать право каталога RBAC
      description: >
        Создаёт пользовательское право (system=false) с произвольной парой
        (action, resource); матчинг строгий. Требует право (manage, iam:global);
        отказ/недоступность IDM → 403 (fail-closed). Дубль пары (action, resource)
        → 409; пустые поля → 400. После мутации IDM широко инвалидирует кэш
        решений (ADR-0015).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePermissionRequest"
      responses:
        "201":
          description: Право создано
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Permission"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: deletePermission
      summary: Удалить право каталога RBAC
      description: >
        Удаляет право по паре (action, resource), переданной в query-параметрах
        (каскадно убирает связки права с ролями). Требует право (manage,
        iam:global); отказ/недоступность IDM → 403 (fail-closed). Системное право
        защищено от удаления → 422; несуществующее право → 404; пустые поля → 400.
        После мутации IDM широко инвалидирует кэш решений (ADR-0015).
      parameters:
        - $ref: "#/components/parameters/ActionQuery"
        - $ref: "#/components/parameters/ResourceQuery"
      responses:
        "200":
          description: Право удалено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Permission"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/PreconditionFailed"
  /iam/roles/{role}/permissions:
    get:
      operationId: getRolePermissions
      summary: Права конкретной роли
      description: >
        Возвращает права роли (read-only). Требует право (read, iam:global);
        отказ/недоступность IDM → 403 (fail-closed); несуществующая роль → 404.
      parameters:
        - $ref: "#/components/parameters/RolePath"
      responses:
        "200":
          description: Права роли
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermissionList"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: attachPermission
      summary: Прикрепить право к роли
      description: >
        Прикрепляет существующее право к роли (идемпотентно: повтор уже
        привязанного → 200). Требует право (manage, iam:global);
        отказ/недоступность IDM → 403 (fail-closed). Несуществующая роль или право
        → 404; системная роль (её состав прав фиксирован) → 422; пустые поля → 400.
        Ответ — актуальный набор прав роли. После мутации IDM широко инвалидирует
        кэш решений, т.к. затронуты все носители роли (ADR-0015).
      parameters:
        - $ref: "#/components/parameters/RolePath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttachPermissionRequest"
      responses:
        "200":
          description: Право прикреплено (или уже было прикреплено)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RolePermissions"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/PreconditionFailed"
    delete:
      operationId: detachPermission
      summary: Открепить право от роли
      description: >
        Открепляет право от роли по паре (action, resource) в query-параметрах
        (идемпотентно: открепление непривязанного → 200). Требует право (manage,
        iam:global); отказ/недоступность IDM → 403 (fail-closed). Несуществующая
        роль → 404; системная роль → 422; пустые поля → 400. Ответ — актуальный
        набор прав роли. После мутации IDM широко инвалидирует кэш решений (ADR-0015).
      parameters:
        - $ref: "#/components/parameters/RolePath"
        - $ref: "#/components/parameters/ActionQuery"
        - $ref: "#/components/parameters/ResourceQuery"
      responses:
        "200":
          description: Право откреплено (или его и не было)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RolePermissions"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/PreconditionFailed"
  /iam/subjects:
    get:
      operationId: listSubjects
      summary: Субъекты с их ролями
      description: >
        Возвращает страницу субъектов (DISTINCT subject из subject_roles) с их
        ролями; keyset-пагинация по subject. Субъекты без ролей системе
        неизвестны и не возвращаются. Требует право (read, iam:global);
        отказ/недоступность IDM → 403 (fail-closed).
      parameters:
        - $ref: "#/components/parameters/PageSizeQuery"
        - $ref: "#/components/parameters/PageTokenQuery"
      responses:
        "200":
          description: Страница субъектов с ролями
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /iam/subjects/{subject}/roles:
    get:
      operationId: getSubjectRoles
      summary: Роли конкретного субъекта
      description: >
        Возвращает роли субъекта (пустой набор, не 404, если ролей нет). Требует
        право (read, iam:global); отказ/недоступность IDM → 403 (fail-closed).
      parameters:
        - $ref: "#/components/parameters/SubjectPath"
      responses:
        "200":
          description: Роли субъекта
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectRoles"
        "403":
          $ref: "#/components/responses/Forbidden"
  /iam/subjects/{subject}/roles/{role}:
    post:
      operationId: assignRole
      summary: Назначить роль субъекту
      description: >
        Назначает субъекту существующую роль (идемпотентно: повтор → 200). Требует
        право (write, iam:global); отказ/недоступность IDM → 403 (fail-closed).
        Несуществующая роль → 404; пустые subject/role → 400. После мутации IDM
        инвалидирует кэш решений по субъекту. Ответ — актуальный набор ролей
        субъекта (ADR-0014).
      parameters:
        - $ref: "#/components/parameters/SubjectPath"
        - $ref: "#/components/parameters/RolePath"
      responses:
        "200":
          description: Роль назначена (или уже была назначена)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectRoles"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: revokeRole
      summary: Снять роль с субъекта
      description: >
        Снимает у субъекта роль (идемпотентно: снятие отсутствующей связки → 200).
        Требует право (write, iam:global); отказ/недоступность IDM → 403
        (fail-closed); пустые subject/role → 400. После мутации IDM инвалидирует
        кэш решений по субъекту. Ответ — актуальный набор ролей субъекта.
      parameters:
        - $ref: "#/components/parameters/SubjectPath"
        - $ref: "#/components/parameters/RolePath"
      responses:
        "200":
          description: Роль снята (или её и не было)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectRoles"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
components:
  parameters:
...
    SubjectPath:
      name: subject
      in: path
      required: true
      description: Идентификатор субъекта (sub из JWT).
      schema:
        type: string
        minLength: 1
    RolePath:
      name: role
      in: path
      required: true
      description: Имя роли каталога RBAC.
      schema:
        type: string
        minLength: 1
    PageSizeQuery:
      name: page_size
      in: query
      required: false
      description: >
        Желаемый размер страницы; сервер клампит к допустимому пределу. Значение
        должно укладываться в int32 (1..2147483647); выход за диапазон — 400.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 2147483647
    PageTokenQuery:
      name: page_token
      in: query
      required: false
      description: >
        Непрозрачный курсор продолжения из предыдущего ответа; пустой — с начала
        выборки.
      schema:
        type: string
    ActionQuery:
      name: action
      in: query
      required: true
      description: Действие права (часть идентификатора пары action/resource).
      schema:
        type: string
        minLength: 1
    ResourceQuery:
      name: resource
      in: query
      required: true
      description: Ресурс права (часть идентификатора пары action/resource).
      schema:
        type: string
        minLength: 1
  responses:
    BadRequest:
      description: Некорректный запрос (валидация входных данных)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Конфликт состояния (например, имя сервиса уже занято)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Доступ запрещён (RBAC, fail-closed)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    PreconditionFailed:
      description: >
        Предусловие операции не выполнено (например, нагрузка не снята из K8s или
        недопустимый исходный статус сервиса).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Role:
      type: object
      required: [name, system]
      properties:
        name:
          type: string
          description: >
            Имя роли — её стабильный идентификатор (им оперируют assign/revoke);
            внутренний id наружу не отдаётся (ADR-0014).
        system:
          type: boolean
          description: >
            Признак системной (сидированной) роли: защищена от удаления и правки
            набора прав через API. UI скрывает для таких роль удаление/правку
            (ADR-0015).
    Permission:
      type: object
      required: [action, resource, system]
      properties:
        action:
          type: string
          description: Действие (например, read, write, create, manage).
        resource:
          type: string
          description: Целевой ресурс (например, iam:global, project:demo).
        system:
          type: boolean
          description: >
            Признак системного (сидированного) права: защищено от удаления через
            API (ADR-0015).
    RolePermissions:
      type: object
      required: [role, permissions]
      properties:
        role:
          type: string
          description: Имя роли.
        permissions:
          type: array
          description: Актуальный набор прав роли после операции.
          items:
            $ref: "#/components/schemas/Permission"
    CreateRoleRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          description: Имя создаваемой роли (непустое, уникальное).
    CreatePermissionRequest:
      type: object
      required: [action, resource]
      properties:
        action:
          type: string
          minLength: 1
          description: Действие создаваемого права.
        resource:
          type: string
          minLength: 1
          description: Ресурс создаваемого права.
    AttachPermissionRequest:
      type: object
      required: [action, resource]
      properties:
        action:
          type: string
          minLength: 1
          description: Действие прикрепляемого права.
        resource:
          type: string
          minLength: 1
          description: Ресурс прикрепляемого права.
    RoleList:
      type: object
      required: [roles]
      properties:
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
    PermissionList:
      type: object
      required: [permissions]
      properties:
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
    SubjectRoles:
      type: object
      required: [subject, roles]
      properties:
        subject:
          type: string
          description: Идентификатор субъекта (sub из JWT).
        roles:
          type: array
          description: Имена ролей субъекта (пусто, если ролей нет).
          items:
            type: string
    SubjectList:
      type: object
      required: [subjects, next_page_token]
      properties:
        subjects:
          type: array
          items:
            $ref: "#/components/schemas/SubjectRoles"
        next_page_token:
          type: string
          description: Курсор следующей страницы; пустой — конец выборки.
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: >
            Стабильное сообщение об ошибке для клиента. Внутренние детали
            (текст gRPC-ошибки) сюда НЕ попадают — только в лог сервера.